refactor fhir-convert
