+++ containers/fhir-converter/Dockerfile
@@ -28,7 +28,7 @@ WORKDIR /app
 COPY requirements.txt .
 RUN pip3 install -r requirements.txt
 
-COPY main.py .
+COPY app/main.py .
 
 EXPOSE 8080
 CMD uvicorn main:api --host 0.0.0.0 --port 8080
